refactor(pubsub): Split pubsub core into reader and writer
